Update: Actualizado estructura de la base de datos.

+++ db/ProyectoHelpDesk.sql
@@ -3,29 +3,29 @@ CREATE DATABASE IF NOT EXISTS HelpDesk;
 USE HelpDesk;
 
 CREATE TABLE IF NOT EXISTS PERMISOS(
-    pk_per_codigo tinyint auto_increment,
+    pk_per_codigo int auto_increment,
     per_tipo varchar(50) not null,
     per_descripcion varchar(120) not null,
     primary key(pk_per_codigo)
 ) ENGINE=InnoDB DEFAULT CHARSET=latin1;
 
 CREATE TABLE IF NOT EXISTS ROLES(
-    pk_rol_codigo tinyint auto_increment,
+    pk_rol_codigo int auto_increment,
     rol_descripcion varchar(80) not null,
-    fk_per_codigo tinyint not null,
+    fk_per_codigo int not null,
     primary key(pk_rol_codigo),
     foreign key(fk_per_codigo) references PERMISOS(pk_per_codigo)
 ) ENGINE=InnoDB DEFAULT CHARSET=latin1;
 
 CREATE TABLE IF NOT EXISTS DEPENDENCIAS(
-    pk_dep_codigo tinyint auto_increment,
+    pk_dep_codigo int auto_increment,
     dep_nombre varchar(80) not null,
     dep_extension varchar(15),
     primary key(pk_dep_codigo)
 ) ENGINE=InnoDB DEFAULT CHARSET=latin1;
 
 CREATE TABLE IF NOT EXISTS USUARIOS(
-    pk_usu_codigo smallint auto_increment,
+    pk_usu_codigo int auto_increment,
     usu_primer_nombre varchar(30) not null,
     usu_segundo_nombre varchar(30),
     usu_primer_apellido varchar(30) not null,
@@ -37,25 +37,25 @@ CREATE TABLE IF NOT EXISTS USUARIOS(
     usu_password varchar(20) not null,
     usu_fecha_creacion datetime not null,
     usu_fecha_modificacion datetime,
-    fk_rol_codigo tinyint not null,
-    fk_dep_codigo tinyint not null,
+    fk_rol_codigo int not null,
+    fk_dep_codigo int not null,
     primary key(pk_usu_codigo),
     foreign key(fk_rol_codigo) references ROLES(pk_rol_codigo),
     foreign key(fk_dep_codigo) references DEPENDENCIAS(pk_dep_codigo)
 ) ENGINE=InnoDB DEFAULT CHARSET=latin1;
 
 CREATE TABLE IF NOT EXISTS TIPOS_EQUIPOS(
-    pk_tip_codigo tinyint auto_increment,
+    pk_tip_codigo int auto_increment,
     tip_nombre varchar(80) not null,
     primary key(pk_tip_codigo)
 ) ENGINE=InnoDB DEFAULT CHARSET=latin1;
 
 CREATE TABLE IF NOT EXISTS EQUIPOS(
-    pk_equ_codigo smallint auto_increment,
+    pk_equ_codigo int auto_increment,
     equ_nombre varchar(80) not null,
     equ_precio double not null,
-    fk_dep_codigo tinyint not null,
-    fk_tip_codigo tinyint not null,
+    fk_dep_codigo int not null,
+    fk_tip_codigo int not null,
     primary key(pk_equ_codigo),
     foreign key(fk_dep_codigo) references DEPENDENCIAS(pk_dep_codigo),
     foreign key(fk_tip_codigo) references TIPOS_EQUIPOS(pk_tip_codigo)
@@ -69,8 +69,8 @@ CREATE TABLE IF NOT EXISTS SERVICIOS(
     ser_estado varchar(20) not null,
     ser_prioridad varchar(10),
     ser_fecha_modificacion datetime,
-    fk_usu_codigo smallint not null,
-    fk_equ_codigo smallint not null,
+    fk_usu_codigo int not null,
+    fk_equ_codigo int not null,
     primary key(pk_ser_ticket),
     foreign key(fk_usu_codigo) references USUARIOS(pk_usu_codigo),
     foreign key(fk_equ_codigo) references EQUIPOS(pk_equ_codigo)
@@ -84,7 +84,7 @@ CREATE TABLE IF NOT EXISTS DETALLES_SERVICIOS(
 ) ENGINE=InnoDB DEFAULT CHARSET=latin1;
 
 CREATE TABLE IF NOT EXISTS TECNICOS(
-    pk_tec_codigo smallint auto_increment,
+    pk_tec_codigo int auto_increment,
     tec_primer_nombre varchar(30) not null,
     tec_segundo_nombre varchar(30),
     tec_primer_apellido varchar(30) not null,
@@ -98,7 +98,7 @@ CREATE TABLE IF NOT EXISTS TECNICOS(
 
 CREATE TABLE IF NOT EXISTS TEC_SERV_DET(
     pk_tsd_codigo int auto_increment,
-    fk_tec_codigo smallint not null,
+    fk_tec_codigo int not null,
     fk_ser_ticket int not null,
     fk_det_codigo int not null,
     primary key(pk_tsd_codigo),
@@ -108,7 +108,7 @@ CREATE TABLE IF NOT EXISTS TEC_SERV_DET(
 ) ENGINE=InnoDB DEFAULT CHARSET=latin1;
 
 CREATE TABLE IF NOT EXISTS POLITICAS(
-    pk_pol_codigo smallint auto_increment,
+    pk_pol_codigo int auto_increment,
     pol_descripcion varchar(120) not null,
     primary key(pk_pol_codigo)
 ) ENGINE=InnoDB DEFAULT CHARSET=latin1;
@@ -122,12 +122,12 @@ CREATE TABLE IF NOT EXISTS HISTORIAL_CAMBIOS(
 ) ENGINE=InnoDB DEFAULT CHARSET=latin1;
 
 CREATE TABLE IF NOT EXISTS PARTES_EQUIPOS(
-    pk_pte_codigo smallint auto_increment,
+    pk_pte_codigo int auto_increment,
     pte_serial varchar(50) not null,
     pte_marca varchar(60) not null,
     pte_descripcion varchar(100),
     pte_estado varchar(20) not null,
-    fk_equ_codigo smallint not null,
+    fk_equ_codigo int not null,
     fk_htc_codigo int not null,
     primary key(pk_pte_codigo),
     foreign key(fk_equ_codigo) references EQUIPOS(pk_equ_codigo),
@@ -135,18 +135,18 @@ CREATE TABLE IF NOT EXISTS PARTES_EQUIPOS(
 ) ENGINE=InnoDB DEFAULT CHARSET=latin1;
 
 CREATE TABLE IF NOT EXISTS GARANTIAS(
-    pk_gar_codigo smallint not null,
+    pk_gar_codigo int not null,
     gar_tipo varchar(20),
     gar_fecha_inicio date not null,
     gar_fecha_fin date not null,
     gar_descripcion varchar(80),
-    fk_equ_codigo smallint not null,
+    fk_equ_codigo int not null,
     primary key(pk_gar_codigo),
     foreign key(fk_equ_codigo) references EQUIPOS(pk_equ_codigo)
 ) ENGINE=InnoDB DEFAULT CHARSET=latin1;
 
 CREATE TABLE IF NOT EXISTS LICENCIAS(
-    pk_lic_codigo smallint auto_increment,
+    pk_lic_codigo int auto_increment,
     lic_nombre varchar(30) not null,
     lic_fecha_compra date not null,
     lic_fecha_caducidad date not null,
@@ -156,9 +156,9 @@ CREATE TABLE IF NOT EXISTS LICENCIAS(
 ) ENGINE=InnoDB DEFAULT CHARSET=latin1;
 
 CREATE TABLE IF NOT EXISTS EQUIPOS_LICENCIAS(
-    pk_eql_codigo smallint auto_increment,
-    fk_equ_codigo smallint not null,
-    fk_lic_codigo smallint not null,
+    pk_eql_codigo int auto_increment,
+    fk_equ_codigo int not null,
+    fk_lic_codigo int not null,
     primary key(pk_eql_codigo),
     foreign key(fk_equ_codigo) references EQUIPOS(pk_equ_codigo),
     foreign key(fk_lic_codigo) references LICENCIAS(pk_lic_codigo)
@@ -199,13 +199,13 @@ CREATE TABLE IF NOT EXISTS FICHAS_TECNICAS(
     fic_lumens varchar(10),
     fic_resolucion varchar(10),
     fic_lente varchar(10),
-    fk_equ_codigo smallint not null,
+    fk_equ_codigo int not null,
     primary key(pk_fic_codigo),
     foreign key(fk_equ_codigo) references EQUIPOS(pk_equ_codigo)
 ) ENGINE=InnoDB DEFAULT CHARSET=latin1;
 
 CREATE TABLE IF NOT EXISTS PROVEEDORES (
-    pk_pro_codigo smallint NOT NULL,
+    pk_pro_codigo int NOT NULL,
     pro_nombre varchar(60) NOT NULL,
     pro_telefono varchar(20) NOT NULL,
     pro_direccion varchar(60),
@@ -216,8 +216,8 @@ CREATE TABLE IF NOT EXISTS PROVEEDORES (
 
 CREATE TABLE IF NOT EXISTS PROVEEDORES_EQUIPOS (
     pk_pve_codigo int auto_increment,
-    fk_equ_codigo smallint NOT NULL,
-    fk_pro_codigo smallint NOT NULL,
+    fk_equ_codigo int NOT NULL,
+    fk_pro_codigo int NOT NULL,
     PRIMARY KEY (pk_pve_codigo),
     FOREIGN KEY (fk_equ_codigo) REFERENCES EQUIPOS(pk_equ_codigo),
     FOREIGN KEY (fk_pro_codigo) REFERENCES PROVEEDORES(pk_pro_codigo)
@@ -225,8 +225,8 @@ CREATE TABLE IF NOT EXISTS PROVEEDORES_EQUIPOS (
 
 CREATE TABLE IF NOT EXISTS PARTES_PROVEEDORES (
     pk_ppv_codigo int auto_increment,
-    fk_pte_codigo smallint,
-    fk_pro_codigo smallint NOT NULL,
+    fk_pte_codigo int,
+    fk_pro_codigo int NOT NULL,
     PRIMARY KEY (pk_ppv_codigo),
     FOREIGN KEY (fk_pte_codigo) REFERENCES PARTES_EQUIPOS(pk_pte_codigo),
     FOREIGN KEY (fk_pro_codigo) REFERENCES PROVEEDORES(pk_pro_codigo)
